requirements.txt
ascii-magic==2.3.0
beautifulsoup4==4.14.3
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==12.0.0
pycryptodome==3.23.0
Pygments==2.19.2
python-dotenv==1.1.1
qrcode==8.2
requests==2.32.5
rich==13.9.4
soupsieve==2.8.1
typing_extensions==4.15.0
urllib3==2.5.0
